+++ src/mc_control/samples/FSM/src/mc_fsm_controller.h
@@ -16,6 +16,11 @@ struct MC_CONTROL_DLLAPI FSMController : public mc_control::fsm::Controller
 
   void reset(const mc_control::ControllerResetData & reset_data) override;
 
+  void supported_robots(std::vector<std::string> & out) const override
+  {
+    out = {"jvrc1"};
+  }
+
 private:
   mc_rtc::Configuration config_;
 };