+++ src/mc_control/samples/LIPMStabilizer/src/mc_lipm_stabilizer.h
@@ -16,6 +16,11 @@ struct MC_CONTROL_DLLAPI LIPMStabilizerController : public mc_control::fsm::Cont
 
   void reset(const mc_control::ControllerResetData & reset_data) override;
 
+  void supported_robots(std::vector<std::string> & out) const override
+  {
+    out = {"jvrc1", "hrp2_drc", "hrp4", "hrp5_p"};
+  }
+
 private:
   mc_rtc::Configuration config_;
 };